Adding UI to State System * need Fix *

+++ Assets/Scripts/Health/IHaveHealth.cs
@@ -6,7 +6,7 @@ using UnityEngine.PlayerLoop;
 
 public class IHaveHealth : MonoBehaviour
 {
-    public int currentHealth;
+    public float currentHealth;
 
     void Update()
     {
@@ -16,7 +16,7 @@ public class IHaveHealth : MonoBehaviour
         }
     }
 
-    public void SetHealth(int _MaxHealth)
+    public void SetHealth(float _MaxHealth)
     {
         currentHealth = _MaxHealth;
     }
